feat: Mouth obscuration is now config dependant (#25)

+++ code/modules/reagents/reagent_containers/pill.dm
@@ -24,6 +24,8 @@
 /obj/item/reagent_containers/food/pill/attack(mob/living/carbon/M, mob/user, def_zone)
 	if(!istype(M))
 		return FALSE
+	if(M.is_mouth_obscured(user))
+		return FALSE
 	bitesize = reagents.total_volume
 	if(M.eat(src, user))
 		qdel(src)
